feat(ds): relicense EmptyState to Apache-2.0 as shared/ds/Empty (#443, D-021)
The design system had no empty state and could not have one: shared/ds is
Apache-2.0 (D-014, D-016) and may not depend on or re-export the AGPL
core, while src/shared/EmptyState.tsx was AGPL with 24 importers and a
documented history of three divergent empty states being consolidated
into it. D-021 chose the relicence over building a second generic Empty
beside it, because that is the only option leaving the product with ONE
empty state and a complete primitive layer.

The implementation moves to shared/ds/Empty.tsx under Apache-2.0 and
shared/EmptyState.tsx becomes a re-export shim under the original names,
so none of the 24 call sites change — Résolution 1 point 4 forbids
rewriting them.

softFill is inlined as a local one-line color-mix rather than imported
from shared/riskColors: an Apache file may not depend on the AGPL core,
and there is nothing to share — the rest of riskColors maps criticality
and framework names, which is product knowledge that stays AGPL.
